Updates to bufferMESSpositives function
Function now updated to include the following:
- Suppress gBuffer warning
- Clip buffered points by the MES surface (only areas of interpolation
are kept). To achieve this, MESS surface is turned into a shapefile, and
then a clip is performed using gIntersection
- Retain original shapefile attributes (COUNTRY_ID, ADMIN_LEVEL and
Med_Class), once buffered, clipped areas have been merged with the
current EOR.
- Generate a medical class for new buffered points. If the point falls
within a country with an existing classification, it is assigned that
medical class. If the point falls within a new country; class is
assigned based on either (a) majority class; i.e. if all locations are
class 2, the new range is class 2, or (b) assigned medical class 1 for
species which are both C1 and C2